refactor: Phase 3 — MCP tools declared in plugin manifests
Each plugin now owns its MCP tool definitions + handlers via mcpTools
on its manifest. The MCP route uses collectMcpTools() from the kernel
to gather all tools dynamically — no more hardcoded array or switch.

Tool ownership: content (1), analytics (2), discovery (1),
payments (2), telegram (1), mcp (5 — site_info + 4 network tools).
