+++ plugins/analytics/manifest.ts
@@ -1,11 +1,13 @@
 import type { PluginManifest, HonoApp } from '../../kernel/types'
 import { analyticsRoutes } from './routes'
+import { analyticsMcpTools } from './mcp-tools'
 
 const analyticsPlugin: PluginManifest = {
   name: 'analytics',
   version: '1.0.0',
   description: 'Page views, reactions, subscribers, and feedback tracking',
   requiredRole: 'viewer',
+  mcpTools: analyticsMcpTools,
 
   mountRoutes: (app: HonoApp) => {
     app.route('/api', analyticsRoutes)
